Mocked quote WIP...
